requirements.txt
# Web scraping and automation
playwright==1.40.0
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
selenium==4.15.2
fake-useragent==1.4.0

# NLP and sentiment analysis
vaderSentiment==3.3.2
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Data processing
python-dateutil==2.8.2
pytz==2023.3
jsonschema==4.19.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
coloredlogs==15.0

# GitHub API
PyGithub==2.1.1

# Stealth scraping
playwright-stealth==1.0.0
undetected-chromedriver==3.5.3

# For finBERT model
transformers[torch]
torchvision

# Additional utilities
tqdm==4.66.1
schedule==1.2.0